Move mountain selection on Create Peak List page to seperate modal
